+++ src/Exception/InvalidArgumentException.php
@@ -2,7 +2,11 @@
 
 namespace Solital\Exception;
 
-class InvalidArgumentException extends \Exception
+class InvalidArgumentException
 {
-    
+    public static function alert(int $code, string $type, string $msg)
+    {
+        include_once "template/error-http.php";
+        die;
+    }
 }